+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
+## Dev
+
+- fixed passing injected parameters
+
 ## 2.0.2
 
 - fixed default container
+- container can be injected
 
 ## 2.0.1